feat(test): add #testState case
the test case test to no forget state of a rule
